Comments on the rest of the controllers

+++ discgolfapp/backend/src/controllers/userController.js
@@ -2,14 +2,16 @@ import * as userService from '../services/userService.js'
 
 
 /**
- * @author Lars263506 (Github)
+ * @author Lars Andreas Strand
  * @description This controller contains request handlers for user registration, authentication and user data management
  */
 
 /**
- * @returns User array, excluding hashed password
- * @description Gets all users from the database for testing and debugging purposes
- * @throws Error if no users were found
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get all users.
+ * It retrieves all users from the database and sends them as a response.
+ * If successful, it sends a 200 status code and the users data.
+ * If there is an error, it sends a 404 status code and the error message.
  */
 
 const getAllUsers = async (req, res) => {
@@ -22,10 +24,11 @@ const getAllUsers = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Checks the permissions of a user
- * @throws Error if the user was not found
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get permissions for a user.
+ * It retrieves the permissions from the database and sends them as a response.
+ * If successful, it sends a 200 status code and the permissions data.
+ * If there is an error, it sends a 404 status code and the error message.
  */
 
 const getPermissions = async (req, res) => {
@@ -37,10 +40,11 @@ const getPermissions = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Gets the profile of a user
- * @throws Error if the user was not found
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get the profile of a user.
+ * It retrieves the profile from the database and sends it as a response.
+ * If successful, it sends a 200 status code and the profile data except hashed password.
+ * If there is an error, it sends a 404 status code and the error message.
  */
 
 const getProfile = async (req, res) => {
@@ -52,10 +56,11 @@ const getProfile = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Gets the profile of a user
- * @throws Error if the user was not found
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get the profile image of a user.
+ * It retrieves the image from the database and sends it as a response.
+ * If successful, it sends a 200 status code and the image data.
+ * If there is an error, it sends a 404 status code and the error message.
  */
 
 const getProfileImage = async (req, res) => {
@@ -67,10 +72,11 @@ const getProfileImage = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Gets a user from the database by id
- * @throws Error if the user was not found
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get a specific user by ID.
+ * It retrieves the user from the database and sends it as a response.
+ * If successful, it sends a 200 status code and the user data (except hashed password).
+ * If there is an error, it sends a 404 status code and the error message.
  */
 
 const getUser = async (req, res) => {
@@ -84,10 +90,11 @@ const getUser = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Gets a user from the database by email
- * @throws Error if the user was not found
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get a specific user by email.
+ * It retrieves the user from the database and sends it as a response.
+ * If successful, it sends a 200 status code and the user data (except hashed password).
+ * If there is an error, it sends a 404 status code and the error message.
  */
 
 const getUserByEmail = async (req, res) => {
@@ -100,15 +107,31 @@ const getUserByEmail = async (req, res) => {
   }
 }
 
+/**
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get all clubs the user is a member of.
+ * It retrieves all clubs for the user from the database and sends them as a response.
+ * If successful, it sends a 200 status code and the clubs data.
+ * If there is an error, it sends a 404 status code and the error message.
+ */
+
 const getUserClubs = async (req, res) => {
   try {
     const userId = req.user.id
     const clubs = await userService.getUserClubs(userId)
-    res.status(200).json(clubs);  
+    res.status(200).json(clubs);
   } catch (error) {
-    res.status(500).json({ error: error.message }); 
+    res.status(500).json({ error: error.message });
   }
 }
+
+/**
+ * @author Lars Andreas Strand
+ * @description This function handles the request to get all games the user has played.
+ * It retrieves all games for the user from the database and sends them as a response.
+ * If successful, it sends a 200 status code and the games data.
+ * If there is an error, it sends a 404 status code and the error message.
+ */
 
 const getUserGames = async (req, res) => {
   try {
@@ -121,8 +144,11 @@ const getUserGames = async (req, res) => {
 }
 
 /**
- * @returns boolean that is true if the user is an admin, false if not
- * @description Checks if a user is an admin
+ * @author Lars Andreas Strand
+ * @description This function handles the request to check if a user is an admin.
+ * It checks the user's role and sends a response indicating if the user is an admin or not.
+ * If successful, it sends a 200 status code and the isAdmin status.
+ * If there is an error, it sends a 500 status code and the error message.
  */
 
 const checkIfAdmin = async (req, res) => {
@@ -130,15 +156,16 @@ const checkIfAdmin = async (req, res) => {
     const isAdmin = req.user.role === 'admin'
     res.status(200).json({ isAdmin })
   } catch (error) {
-    res.status(500).json({ error: "Couldn't check if user is an admin, try again later." }); 
+    res.status(500).json({ error: "Couldn't check if user is an admin, try again later." });
   }
 }
 
 /**
- * @param req
- * @param res
- * @description Registers a new user in the database
- * @throws Error if there was an error registering the user in the database
+ * @author Lars Andreas Strand
+ * @description This function handles the request to register a new user.
+ * It retrieves the user data from the request body and calls the service to register the user.
+ * If successful, it sends a 201 status code and the created user data.
+ * If there is an error, it sends a 400 status code and the error message.
  */
 
 const registerUser = async (req, res) => {
@@ -152,10 +179,11 @@ const registerUser = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Logs in a user and tokens are created for the user
- * @throws Error if the email or password is incorrect
+ * @author Lars Andreas Strand
+ * @description This function handles the request to log in a user.
+ * It retrieves the email and password from the request body and calls the service to log in the user.
+ * If successful, it sends a 200 status code and the access token, refresh token and display name.
+ * If there is an error, it sends a 401 status code and the error message.
  */
 
 const loginUser = async (req, res) => {
@@ -169,10 +197,11 @@ const loginUser = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Logs in a user and tokens are created for the user
- * @throws Error if the email or password is incorrect
+ * @author Lars Andreas Strand
+ * @description This function handles the request to upload a profile image for a user.
+ * It retrieves the user ID from the request and the image data from the request file.
+ * If successful, it sends a 200 status code and the result of the image upload.
+ * If there is an error, it sends a 500 status code and the error message.
  */
 
 const postProfileImage = async (req, res) => {
@@ -186,31 +215,45 @@ const postProfileImage = async (req, res) => {
 };
 
 /**
- * @param req
- * @param res
- * @description Changes the display name of a user in the database
- * @throws Error if there was an error changing the display name in the database
+ * @author Lars Andreas Strand
+ * @description This function handles the request to change the display name of a user.
+ * It retrieves the email and new display name from the request body and calls the service to change the display name.
+ * If successful, it sends a 200 status code and the updated display name.
+ * If there is an error, it sends a 404 status code and the error message.
+ * If the email or new display name is missing, it sends a 400 status code and an error message.
  */
 
 const changeDisplayName = async (req, res) => {
   const { email, newDisplayName } = req.body
+
+  if (!email || !newDisplayName) {
+    return res.status(400).json({ message: 'Email and new display name are required' })
+  }
+
   try {
     const { emailChangedAt } = await userService.changeDisplayName(email, newDisplayName)
     res.status(200).json(emailChangedAt)
   } catch (error) {
-    res.status(400).json({ message: error.message })
+    res.status(404).json({ message: error.message })
   }
 }
 
 /**
- * @param req
- * @param res
- * @description Changes the email of a user in the database
- * @throws Error if there was an error changing the email in the database
+ * @author Lars Andreas Strand
+ * @description This function handles the request to change the email of a user.
+ * It retrieves the email and new email from the request body and calls the service to change the email.
+ * If successful, it sends a 200 status code and the updated email.
+ * If there is an error, it sends a 404 status code and the error message.
+ * If the email or new email is missing, it sends a 400 status code and an error message.
  */
 
 const changeEmail = async (req, res) => {
   const { email, newEmail } = req.body
+
+  if (!email || !newEmail) {
+    return res.status(400).json({ message: 'Email and new email are required' })
+  }
+
   try {
     const { emailChangedAt } = await userService.changeEmail(email, newEmail)
     res.status(200).json(emailChangedAt)
@@ -220,53 +263,74 @@ const changeEmail = async (req, res) => {
 }
 
 /**
- * @param req
- * @param res
- * @description Changes the password of a user in the database
- * @throws Error if there was an error changing the password in the database
+ * @author Lars Andreas Strand
+ * @description This function handles the request to change the password of a user.
+ * It retrieves the email and new password from the request body and calls the service to change the password.
+ * If successful, it sends a 200 status code and the updated password.
+ * If there is an error, it sends a 404 status code and the error message.
+ * If the email or new password is missing, it sends a 400 status code and an error message.
  */
 
 const changePassword = async (req, res) => {
   const { email, newPassword } = req.body
+
+  if (!email || !newPassword) {
+    return res.status(400).json({ message: 'Email and new password are required' })
+  }
+
   try {
     const { passwordChangedAt } = await userService.changePassword(email, newPassword)
     res.status(200).json(passwordChangedAt)
   } catch (error) {
-    res.status(400).json({ message: error.message })
+    res.status(404).json({ message: error.message })
   }
 }
 
 /**
- * @param req
- * @param res
- * @description Changes the role of a user in the database
- * @throws Error if there was an error changing the role in the database
+ * @author Lars Andreas Strand
+ * @description This function handles the request to change the role of a user.
+ * It retrieves the email and new role from the request body and calls the service to change the role.
+ * If successful, it sends a 200 status code and the updated role.
+ * If there is an error, it sends a 404 status code and the error message.
+ * If the email or new role is missing, it sends a 400 status code and an error message.
  */
 
 const changeRole = async (req, res) => {
   const { email, newRole } = req.body
+
+  if (!email || !newRole) {
+    return res.status(400).json({ message: 'Email and new role are required' })
+  }
+
   try {
     const { userRole, roleChangedAt } = await userService.changeRole(email, newRole)
     res.status(200).json({ userRole, roleChangedAt })
   } catch (error) {
-    res.status(400).json({ message: error.message })
+    res.status(404).json({ message: error.message })
   }
 }
 
 /**
- * @param req
- * @param res
- * @description Deletes a user from the database
- * @throws Error if there was an error deleting the user from the database
+ * @author Lars Andreas Strand
+ * @description This function handles the request to delete a user.
+ * It retrieves the email from the request body and calls the service to delete the user.
+ * If successful, it sends a 200 status code and the deleted user data.
+ * If there is an error, it sends a 404 status code and the error message.
+ * If the email is missing, it sends a 400 status code and an error message.
  */
 
 const deleteUser = async (req, res) => {
   const { email } = req.body
+
+  if (!email) {
+    return res.status(400).json({ message: 'Email is required' })
+  }
+
   try {
     const user = await userService.deleteUser(email)
     res.status(200).json(user)
   } catch (error) {
-    res.status(400).json({ message: error.message })
+    res.status(404).json({ message: error.message })
   }
 }
